Factor out the reflective connection factory impl.
It didn't really belong where it was.
